Fix bugs

+++ Java Playgrnd/Java Playgrnd Assembler/jplasm.c
@@ -2,7 +2,7 @@
 #include "Java Playgrnd Assembler.h"
 
 
-int main(size_t argc, char** argv)
+int main(int argc, char** argv)
   {
     assemble(argv[1], argv[2]);
